refactor: rename StudentEventsList folder to EventList and delete previous EventsList
