--- src/components/Events/EventsList/EventList-styles.less
@@ -1,7 +1,0 @@
-.imageList {
-	border-radius: 10px;
-}
-
-.event-item {
-	margin: 0px;
-}
