+++ src/components/Events/index.ts
@@ -1,3 +1,2 @@
 export { EventDetail } from './EventDetail/EventDetail'
 export { EventForm } from './EventForm/EventForm'
-export { EventList } from './EventsList/EventList'
