+++ src/components/StudentSection/EventList/StudentEventsList-styles.less
