+++ src/components/StudentSection/EventList/_components/EventCard/EventCard-types.tsx
