+++ src/components/StudentSection/EventList/_components/EventCard/EventCard-utils.tsx
