+++ src/components/StudentSection/EventList/_components/EventCard/EventCard.tsx
