+++ src/components/StudentSection/EventList/_components/Skeleton/Skeleton-styles.less
