+++ src/components/index.ts
@@ -7,7 +7,7 @@ export { ContactsLayout } from './ContactsLayout/ContactsLayout'
 export { ContactsMap } from './ContactsMap/ContactsMap'
 export { CopyableText } from './CopyableText/CopyableText'
 export { Divider } from './Divider/Divider'
-export { EventForm, EventDetail, EventList } from './Events'
+export { EventForm, EventDetail } from './Events'
 export { StudentEventList, StudentEventDetail, Reward } from './StudentSection'
 export { FavoritesList } from './FavoritesList/FavoritesList'
 export { HomeLayout } from './HomeLayout/HomeLayout'
